Update quiz commands
Only owner or admin can edit quizzes

+++ src/Application/Quizzes/Commands/DeleteQuiz/DeleteQuizCommand.cs
@@ -13,21 +13,29 @@ public class DeleteQuizCommand : IRequest
     {
         private readonly IApplicationDbContext _context;
         private readonly ISFXFileBuilder _SFXFileBulider;
+        private readonly ICurrentUserService _currentUserService;
+        private readonly IIdentityService _identityService;
 
-        public DeleteQuizCommandHandler(IApplicationDbContext context, ISFXFileBuilder sFXFileBulider)
+        public DeleteQuizCommandHandler(IApplicationDbContext context, ISFXFileBuilder sFXFileBulider, ICurrentUserService currentUserService, IIdentityService identityService)
         {
             _context = context;
             _SFXFileBulider = sFXFileBulider;
+            _currentUserService = currentUserService;
+            _identityService = identityService;
         }
 
         public async Task<Unit> Handle(DeleteQuizCommand request, CancellationToken cancellationToken)
         {
-            //TODO: delete if quiz owner or is in role admin
-
             Quiz entity = await _context.Quizzes.FindAsync(request.Id);
 
             if (entity is null)
                 throw new NotFoundException(nameof(Quiz), request.Id);
+
+
+            bool isInRoleAdmin = await _identityService.IsInRoleAsync(_currentUserService.UserId, "Administrator");
+
+            if (!(entity.CreatedBy == _currentUserService.UserId) && !isInRoleAdmin)
+                throw new ForbiddenAccessException();
 
             _context.Quizzes.Remove(entity);
